fix: correct the native identity layout and durable undo receipts
The FILE_ID_INFO interop declared the volume serial as 32-bit DWORD
where the native struct uses ULONGLONG - the undersized 20-byte buffer
made every GetFileInformationByHandleEx(FileIdInfo) call fail, so the
128-bit upgrade silently never ran and identity kept running on the
legacy 64-bit view. The layout is now correct (24 bytes), the volume
serial is ulong throughout, and the legacy 64-bit fallback is deleted:
FileIdInfo is supported on every platform DeskBox runs on, so a failed
query denies identity authority instead of degrading to the index that
is not unique on ReFS.

A plain-organization undo retry after a partial failure now plans only
the items not yet restored - previously every item re-entered the plan,
so the retry either aborted on the already-restored item's vanished
target or relocated whatever reappeared at that path. The plan and its
history item are paired explicitly, so completion bookkeeping cannot
misalign after filtering.

Undo receipts are durable in history: organization history items carry
the destination identity captured at move time, and undo verifies
against that recorded receipt instead of capturing a fresh identity at
undo time - a replacement that later occupies the destination path no
longer receives a valid id from DeskBox itself. Legacy history entries
without an identity keep the pre-existing capture-on-undo behavior.

+++ src/DeskBox/Models/DesktopOrganizationModels.cs
@@ -248,7 +248,7 @@ public sealed class DesktopOrganizationRecoveryJournal
 /// </summary>
 public sealed class DesktopOrganizationDestinationIdentity
 {
-    public uint VolumeSerialNumber { get; set; }
+    public ulong VolumeSerialNumber { get; set; }
 
     public ulong FileIdHigh { get; set; }
